chore: local files storage

+++ src/main/utils/localFilesStorage.ts
@@ -1,9 +1,22 @@
 import fs from 'node:fs';
 import { ensureDir } from './io';
 import { getUserFilesStoragePath } from './customStoragePath';
-import { RevezoneFileTree, RevezoneFolder, RevezoneFile } from '../../renderer/src/types/file';
-import { join } from 'node:path';
+import {
+  RevezoneFileTree,
+  RevezoneFolder,
+  RevezoneFile,
+  RevezoneFileType
+} from '../../renderer/src/types/file';
+import path, { join } from 'node:path';
 import { TreeItem } from 'react-complex-tree';
+
+interface FullPathInfo {
+  type: 'folder' | 'file';
+  fileType?: RevezoneFileType;
+  suffix?: '.excalidraw' | '.md';
+  path: string;
+  parentDirPath: string;
+}
 
 /**
  * ATTENTION: Files's name cannot be same in one directory
@@ -11,7 +24,11 @@ import { TreeItem } from 'react-complex-tree';
  * @param fileTree
  * @param filePath
  */
-export const getFilePath = (itemId: string, fileTree: RevezoneFileTree, filePath = '') => {
+export const getParentPathInFileTree = (
+  itemId: string,
+  fileTree: RevezoneFileTree,
+  filePath = ''
+) => {
   let parentItem;
 
   const items = Object.values(fileTree);
@@ -24,33 +41,79 @@ export const getFilePath = (itemId: string, fileTree: RevezoneFileTree, filePath
 
   if (parentItem) {
     // @ts-ignore
-    filePath = getFilePath(parentItem.id, fileTree, `${parentItem.name}/${filePath}`);
+    filePath = getParentPathInFileTree(parentItem.id, fileTree, `${parentItem.name}/${filePath}`);
   }
 
-  return filePath ? `${filePath}/` : '';
+  return filePath;
 };
 
-export function onFileDataChange(fileId: string, value: string, fileTree: RevezoneFileTree) {
-  const file = fileTree[fileId].data;
+export function getFullPathInfo(itemId: string, fileTree: RevezoneFileTree): FullPathInfo {
+  const item = fileTree[itemId].data;
 
   const userFilesStoragePath = getUserFilesStoragePath();
 
-  console.log('--- fileDataChange ---', fileId, file, userFilesStoragePath);
+  const parentPathInFileTree = getParentPathInFileTree(itemId, fileTree);
 
-  const filePathInFileTree = getFilePath(fileId, fileTree);
+  console.log('--- parentPathInFileTree ---', parentPathInFileTree);
 
-  const fileDir = join(userFilesStoragePath, filePathInFileTree);
+  const parentDirPath = join(userFilesStoragePath, parentPathInFileTree);
 
-  ensureDir(fileDir);
+  if (item.type === 'folder') {
+    const folderFullPath = path.join(parentDirPath, item.name);
 
-  const suffix = file.type === 'board' ? '.excalidraw' : '.md';
-  const fullFilePath = `${fileDir}/${file.name}${suffix}`;
+    return {
+      type: 'folder',
+      path: folderFullPath,
+      parentDirPath
+    };
+  } else {
+    const suffix = item.type === 'board' ? '.excalidraw' : '.md';
+    const fullFilePath = path.join(parentDirPath, `${item.name}${suffix}`);
+
+    return {
+      type: 'file',
+      fileType: item.type,
+      suffix,
+      path: fullFilePath,
+      parentDirPath
+    };
+  }
+}
+
+export function addOrUpdateFile(
+  fileId: string,
+  value: string,
+  fileTree: RevezoneFileTree,
+  type: 'add' | 'update'
+) {
+  const { path: fullFilePath, parentDirPath } = getFullPathInfo(fileId, fileTree);
+
+  console.log('--- addOrUpdateFile ---', fullFilePath, type);
+
+  if (type === 'update' && !fs.existsSync(fullFilePath)) {
+    return;
+  }
+
+  ensureDir(parentDirPath);
 
   fs.writeFileSync(fullFilePath, value);
 }
 
-export function onRenameFileOrFolder(fileId: string, newName: string, fileTree: RevezoneFileTree) {
-  console.log('--- rename ---', fileId, newName, fileTree);
+export function onAddFile(fileId: string, value: string, fileTree: RevezoneFileTree) {
+  console.log('--- onAddFile ---', fileId, value, fileTree);
+  addOrUpdateFile(fileId, value, fileTree, 'add');
+}
+
+export function onFileDataChange(fileId: string, value: string, fileTree: RevezoneFileTree) {
+  addOrUpdateFile(fileId, value, fileTree, 'update');
+}
+
+export function onRenameFileOrFolder(itemId: string, newName: string, fileTree: RevezoneFileTree) {
+  console.log('--- rename ---', itemId, newName, fileTree);
+
+  const { path: fullFilePath, parentDirPath, suffix } = getFullPathInfo(itemId, fileTree);
+
+  fs.renameSync(fullFilePath, `${parentDirPath}/${newName}${suffix}`);
 }
 
 export function onDeleteFileOrFolder(
@@ -59,19 +122,11 @@ export function onDeleteFileOrFolder(
 ) {
   console.log('--- delete ---', item, fileTree);
 
-  const filePathInFileTree = getFilePath(item.id, fileTree);
-
-  console.log('--- filePathInFileTree ---', filePathInFileTree);
-
-  const userFilesStoragePath = getUserFilesStoragePath();
-  const fileDir = join(userFilesStoragePath, filePathInFileTree);
+  const { path: fullFilePath } = getFullPathInfo(item.id, fileTree);
 
   if (item.type === 'folder') {
-    fs.rmdirSync(fileDir);
+    fs.rmdirSync(fullFilePath);
   } else {
-    const suffix = item.type === 'board' ? '.excalidraw' : '.md';
-    const fullFilePath = join(fileDir, `${item.name}${suffix}`);
-
     console.log('--- delete file path ---', fullFilePath);
 
     fs.rmSync(fullFilePath);
